bugfix single domain redirect
I came across of this problem with the following domains:

domain | context
----|----
jens.de | web (default)
jens.de/de/ | web
jens.eu | en
jens.eu/en/ | en

The Domain jens.eu redirects to the default `web` context, but have to redirect to `en` in this situation. So i think the problem is a missing match. For my project this solution solved it.